feat: add selectors with path argument for collections
